+++ src/editor/core/draw/Draw.ts
@@ -103,6 +103,8 @@ import { LineBreakParticle } from './particle/LineBreakParticle'
 import { MouseObserver } from '../observer/MouseObserver'
 import { LineNumber } from './frame/LineNumber'
 import { PageBorder } from './frame/PageBorder'
+import { ITd } from '../../interface/table/Td'
+import { Actuator } from '../actuator/Actuator'
 
 export class Draw {
   private container: HTMLDivElement
@@ -248,6 +250,7 @@ export class Draw {
     this.globalEvent.register()
 
     this.workerManager = new WorkerManager(this)
+    new Actuator(this)
 
     const { letterClass } = options
     this.LETTER_REG = new RegExp(`[${letterClass.join('')}]`)
@@ -314,6 +317,8 @@ export class Draw {
 
   public isReadonly() {
     switch (this.mode) {
+      case EditorMode.DESIGN:
+        return false
       case EditorMode.READONLY:
       case EditorMode.PRINT:
         return true
@@ -325,8 +330,11 @@ export class Draw {
   }
 
   public isDisabled() {
+    if (this.mode === EditorMode.DESIGN) return false
     const { startIndex, endIndex } = this.range.getRange()
     const elementList = this.getElementList()
+    // 优先判断表格单元格
+    if (this.getTd()?.disabled) return true
     if (startIndex === endIndex) {
       const startElement = elementList[startIndex]
       const nextElement = elementList[startIndex + 1]
@@ -339,6 +347,10 @@ export class Draw {
     return selectionElementList.some(
       element => element.title?.disabled || element.control?.disabled
     )
+  }
+
+  public isDesignMode() {
+    return this.mode === EditorMode.DESIGN
   }
 
   public getOriginalWidth(): number {
@@ -618,6 +630,16 @@ export class Draw {
     return this.footer.getElementList()
   }
 
+  public getTd(): ITd | null {
+    const positionContext = this.position.getPositionContext()
+    const { index, trIndex, tdIndex, isTable } = positionContext
+    if (isTable) {
+      const elementList = this.getOriginalElementList()
+      return elementList[index!].trList![trIndex!].tdList[tdIndex!]
+    }
+    return null
+  }
+
   public insertElementList(payload: IElement[]) {
     if (!payload.length || !this.range.getIsCanInput()) return
     const { startIndex, endIndex } = this.range.getRange()
@@ -698,6 +720,7 @@ export class Draw {
     deleteCount: number,
     ...items: IElement[]
   ) {
+    const isDesignMode = this.isDesignMode()
     if (deleteCount > 0) {
       // 当最后元素与开始元素列表信息不一致时：清除当前列表信息
       const endIndex = start + deleteCount
@@ -724,12 +747,15 @@ export class Draw {
       }
       // 元素删除（不可删除控件忽略）
       if (!this.control.getActiveControl()) {
+        const tdDeletable = this.getTd()?.deletable
         let deleteIndex = endIndex - 1
         while (deleteIndex >= start) {
           const deleteElement = elementList[deleteIndex]
           if (
-            deleteElement?.control?.deletable !== false &&
-            deleteElement?.title?.deletable !== false
+            isDesignMode ||
+            (tdDeletable !== false &&
+              deleteElement?.control?.deletable !== false &&
+              deleteElement?.title?.deletable !== false)
           ) {
             elementList.splice(deleteIndex, 1)
           }
@@ -1795,13 +1821,23 @@ export class Draw {
     ctx: CanvasRenderingContext2D,
     payload: IDrawRowPayload
   ) {
+    const {
+      control: { activeBackgroundColor }
+    } = this.options
     const { rowList, positionList } = payload
+    const activeControlElement = this.control.getActiveControl()?.getElement()
     for (let i = 0; i < rowList.length; i++) {
       const curRow = rowList[i]
       for (let j = 0; j < curRow.elementList.length; j++) {
         const element = curRow.elementList[j]
         const preElement = curRow.elementList[j - 1]
-        if (element.highlight) {
+        if (
+          element.highlight ||
+          (activeBackgroundColor &&
+            activeControlElement &&
+            element.controlId === activeControlElement.controlId &&
+            !this.control.getIsRangeInPostfix())
+        ) {
           // 高亮元素相连需立即绘制，并记录下一元素坐标
           if (
             preElement &&
@@ -1816,13 +1852,15 @@ export class Draw {
               leftTop: [x, y]
             }
           } = positionList[curRow.startIndex + j]
+          // 元素向左偏移量
+          const offsetX = element.left || 0
           this.highlight.recordFillInfo(
             ctx,
-            x,
+            x - offsetX,
             y,
-            element.metrics.width,
+            element.metrics.width + offsetX,
             curRow.height,
-            element.highlight
+            element.highlight || activeBackgroundColor
           )
         } else if (preElement?.highlight) {
           // 之前是高亮元素，当前不是需立即绘制