+++ src/editor/interface/table/Td.ts
@@ -6,6 +6,8 @@ import { IRow } from '../Row'
 export interface ITd {
   conceptId?: string
   id?: string
+  extension?: unknown
+  externalId?: string
   x?: number
   y?: number
   width?: number
@@ -29,4 +31,6 @@ export interface ITd {
   mainHeight?: number // 内容 + 内边距高度
   realHeight?: number // 真实高度（包含跨列）
   realMinHeight?: number // 真实最小高度（包含跨列）
+  disabled?: boolean // 内容不可编辑
+  deletable?: boolean // 内容不可删除
 }