+++ src/main.ts
@@ -1481,6 +1481,10 @@ function initEditorInstance(
     {
       mode: EditorMode.PRINT,
       name: '打印模式'
+    },
+    {
+      mode: EditorMode.DESIGN,
+      name: '设计模式'
     }
   ]
   const modeElement = document.querySelector<HTMLDivElement>('.editor-mode')!